Fork the "empty" prepareStackTrace case for Server builds
