+++ packages/shared/DefaultPrepareStackTrace.js
@@ -1,0 +1,12 @@
+/**
+ * Copyright (c) Meta Platforms, Inc. and affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @flow
+ */
+
+// This is forked in server builds where the default stack frame may be source mapped.
+
+export default ((undefined: any): (Error, CallSite[]) => string);
